Terminando curso de spring boot con documentación usando Swagger

+++ src/main/java/com/sventas/sventas/controller/CustomerController.java
@@ -1,8 +1,11 @@
 package com.sventas.sventas.controller;
 
+import com.sventas.sventas.exception.ModelNotFoundException;
 import com.sventas.sventas.model.Customers;
 import com.sventas.sventas.service.CustomerService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,27 +18,37 @@ public class CustomerController {
     private CustomerService customerService;
 
     @GetMapping
-    public List<Customers> findAll() {
-        return customerService.findAll();
+    public ResponseEntity<List<Customers>> findAll() {
+        return new ResponseEntity<>(customerService.findAll(), HttpStatus.OK);
     }
 
     @PostMapping
-    public  Customers create(@RequestBody Customers customers) {
-        return customerService.create(customers);
+    public  ResponseEntity<Customers> create(@RequestBody Customers customers) {
+        return new ResponseEntity<>(customerService.create(customers), HttpStatus.CREATED);
     }
 
     @PutMapping
-    public Customers update(@RequestBody Customers customers) {
-        return customerService.update(customers);
+    public ResponseEntity<Object> update(@RequestBody Customers customers) {
+        customerService.update(customers);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
-    public Customers findById(@PathVariable("id") Integer client_id) {
-        return customerService.findById(client_id);
+    public ResponseEntity<Customers> findById(@PathVariable("id") Integer client_id) {
+        Customers customers = customerService.findById(client_id);
+        if(customers == null) {
+            throw  new ModelNotFoundException("Cliente no encontrado.");
+        }
+        return new ResponseEntity<>(customers, HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable("id") Integer client_id) {
+    public ResponseEntity<Object> delete(@PathVariable("id") Integer client_id) throws Exception {
+        Customers customers = customerService.findById(client_id);
+        if(customers == null) {
+            throw  new ModelNotFoundException("El cliente que desea eliminar no existe.");
+        }
         customerService.delete(client_id);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 }
